Move cloud related services into separate assemble
