+++ FunderMaps.sln
@@ -11,6 +11,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FunderMaps.Core", "src\Fund
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Laixer.Identity.Dapper", "contrib\Laixer.Identity.Dapper\Laixer.Identity.Dapper.csproj", "{C6F6EC1F-140A-4874-BCBC-A3CE6062D997}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FunderMaps.Cloud", "src\FunderMaps.Cloud\FunderMaps.Cloud.csproj", "{FE4947DD-2A36-4FD6-BC31-674797585837}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -33,6 +35,10 @@ Global
 		{C6F6EC1F-140A-4874-BCBC-A3CE6062D997}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C6F6EC1F-140A-4874-BCBC-A3CE6062D997}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C6F6EC1F-140A-4874-BCBC-A3CE6062D997}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FE4947DD-2A36-4FD6-BC31-674797585837}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FE4947DD-2A36-4FD6-BC31-674797585837}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FE4947DD-2A36-4FD6-BC31-674797585837}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FE4947DD-2A36-4FD6-BC31-674797585837}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
