+++ src/FunderMaps.Cloud/FunderMaps.Cloud.csproj
@@ -1,0 +1,19 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <Authors>Laixer</Authors>
+    <Company>FunderMaps B.V.</Company>
+    <Product>FunderMaps</Product>
+    <Copyright>Copyright Laixer 2019</Copyright>
+    <PackageProjectUrl>https://github.com/Zanhos/FunderMaps</PackageProjectUrl>
+    <RepositoryUrl>https://github.com/Zanhos/FunderMaps</RepositoryUrl>
+    <RepositoryType>Git</RepositoryType>
+    <PackageTags>Foundation, Research, Visualization, Registration</PackageTags>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\FunderMaps.Core\FunderMaps.Core.csproj" />
+  </ItemGroup>
+
+</Project>
